statsdbased/python/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
prometheus-client==0.11.0
statsd-tags==3.2.1.post1
Werkzeug==2.0.1
